wip: fixing graphql scalars - get minimal repro into ci
